--- a/.github/workflows/coverage.yml
+++ b/.github/workflows/coverage.yml
@@ -15,6 +15,6 @@ jobs:
       - name: Setup go
         uses: actions/setup-go@master
         with:
-          go-version: 1.18
+          go-version: 1.19
       - name: Run tests with coverage
         run: ./ci.sh coverage -d "${GITHUB_BASE_REF-HEAD}"