--- a/fuzz_test.go
+++ b/fuzz_test.go
@@ -1,5 +1,4 @@
-//go:build go1.18
-// +build go1.18
+//go:build go1.18 || go1.19
 
 package toml_test