--- a/internal/ast/ast.go
+++ b/internal/ast/ast.go
@@ -11,10 +11,10 @@ import (
 //
 // For example:
 //
-// it := n.Children()
-// for it.Next() {
-// 		it.Node()
-// }
+//	it := n.Children()
+//	for it.Next() {
+//			it.Node()
+//	}
 type Iterator struct {
 	started bool
 	node    *Node