```
goos: linux
goarch: amd64
pkg: github.com/pelletier/go-toml/v2/benchmark
cpu: AMD Ryzen 9 5950X 16-Core Processor           
name                               old time/op    new time/op    delta
UnmarshalDataset/config-2            16.7ms ± 0%    16.4ms ± 0%   -1.92%  (p=0.016 n=4+5)
UnmarshalDataset/canada-2            66.6ms ± 0%    66.9ms ± 1%     ~     (p=0.190 n=4+5)
UnmarshalDataset/citm_catalog-2      21.5ms ± 1%    22.3ms ± 0%   +3.81%  (p=0.008 n=5+5)
UnmarshalDataset/twitter-2           8.41ms ± 1%    8.55ms ± 2%     ~     (p=0.095 n=5+5)
UnmarshalDataset/code-2              74.2ms ± 0%    72.9ms ± 0%   -1.85%  (p=0.008 n=5+5)
UnmarshalDataset/example-2            126µs ± 1%     129µs ± 0%   +2.96%  (p=0.008 n=5+5)
Unmarshal/SimpleDocument/struct-2     476ns ± 1%     481ns ± 1%   +1.10%  (p=0.032 n=5+5)
Unmarshal/SimpleDocument/map-2        672ns ± 1%     682ns ± 0%   +1.39%  (p=0.008 n=5+5)
Unmarshal/ReferenceFile/struct-2     37.5µs ± 1%    37.4µs ± 0%     ~     (p=0.222 n=5+5)
Unmarshal/ReferenceFile/map-2        57.2µs ± 0%    57.0µs ± 0%     ~     (p=0.151 n=5+5)
Unmarshal/HugoFrontMatter-2          10.1µs ± 0%    10.0µs ± 0%   -1.18%  (p=0.008 n=5+5)
Marshal/SimpleDocument/struct-2       325ns ± 0%     324ns ± 0%   -0.25%  (p=0.032 n=5+5)
Marshal/SimpleDocument/map-2          504ns ± 0%     458ns ± 0%   -9.12%  (p=0.008 n=5+5)
Marshal/ReferenceFile/struct-2       29.6µs ± 0%    27.9µs ± 0%   -5.95%  (p=0.008 n=5+5)
Marshal/ReferenceFile/map-2          39.8µs ± 1%    36.7µs ± 0%   -7.91%  (p=0.008 n=5+5)
Marshal/HugoFrontMatter-2            7.14µs ± 0%    6.71µs ± 0%   -6.11%  (p=0.008 n=5+5)
[Geo mean]                           89.8µs         88.4µs        -1.54%

name                               old speed      new speed      delta
UnmarshalDataset/config-2          62.7MB/s ± 0%  63.9MB/s ± 0%   +1.95%  (p=0.016 n=4+5)
UnmarshalDataset/canada-2          33.2MB/s ± 2%  32.9MB/s ± 1%     ~     (p=0.095 n=5+5)
UnmarshalDataset/citm_catalog-2    26.0MB/s ± 1%  25.0MB/s ± 0%   -3.68%  (p=0.008 n=5+5)
UnmarshalDataset/twitter-2         52.6MB/s ± 1%  51.7MB/s ± 2%     ~     (p=0.095 n=5+5)
UnmarshalDataset/code-2            36.2MB/s ± 0%  36.8MB/s ± 0%   +1.88%  (p=0.008 n=5+5)
UnmarshalDataset/example-2         64.5MB/s ± 1%  62.7MB/s ± 0%   -2.88%  (p=0.008 n=5+5)
Unmarshal/SimpleDocument/struct-2  23.1MB/s ± 1%  22.8MB/s ± 1%   -1.09%  (p=0.032 n=5+5)
Unmarshal/SimpleDocument/map-2     16.4MB/s ± 1%  16.1MB/s ± 0%   -1.37%  (p=0.008 n=5+5)
Unmarshal/ReferenceFile/struct-2    140MB/s ± 1%   140MB/s ± 0%     ~     (p=0.222 n=5+5)
Unmarshal/ReferenceFile/map-2      91.6MB/s ± 0%  91.9MB/s ± 0%     ~     (p=0.103 n=5+5)
Unmarshal/HugoFrontMatter-2        54.0MB/s ± 0%  54.7MB/s ± 0%   +1.20%  (p=0.008 n=5+5)
Marshal/SimpleDocument/struct-2    37.0MB/s ± 0%  37.0MB/s ± 0%   +0.25%  (p=0.032 n=5+5)
Marshal/SimpleDocument/map-2       23.8MB/s ± 0%  26.2MB/s ± 0%  +10.04%  (p=0.008 n=5+5)
Marshal/ReferenceFile/struct-2     69.5MB/s ± 0%  73.9MB/s ± 0%   +6.32%  (p=0.008 n=5+5)
Marshal/ReferenceFile/map-2        50.3MB/s ± 1%  54.7MB/s ± 0%   +8.58%  (p=0.008 n=5+5)
Marshal/HugoFrontMatter-2          73.2MB/s ± 0%  78.0MB/s ± 0%   +6.50%  (p=0.008 n=5+5)
[Geo mean]                         45.9MB/s       46.6MB/s        +1.54%

name                               old alloc/op   new alloc/op   delta
UnmarshalDataset/config-2            5.86MB ± 0%    5.86MB ± 0%     ~     (p=1.000 n=5+5)
UnmarshalDataset/canada-2            83.0MB ± 0%    83.0MB ± 0%     ~     (p=0.667 n=4+5)
UnmarshalDataset/citm_catalog-2      34.7MB ± 0%    34.7MB ± 0%     ~     (p=0.841 n=5+5)
UnmarshalDataset/twitter-2           12.7MB ± 0%    12.7MB ± 0%   -0.07%  (p=0.008 n=5+5)
UnmarshalDataset/code-2              22.2MB ± 0%    22.2MB ± 0%     ~     (p=0.968 n=5+5)
UnmarshalDataset/example-2            186kB ± 0%     186kB ± 0%   -0.01%  (p=0.024 n=5+5)
Unmarshal/SimpleDocument/struct-2      805B ± 0%      805B ± 0%     ~     (all equal)
Unmarshal/SimpleDocument/map-2       1.13kB ± 0%    1.13kB ± 0%     ~     (all equal)
Unmarshal/ReferenceFile/struct-2     20.9kB ± 0%    20.9kB ± 0%     ~     (all equal)
Unmarshal/ReferenceFile/map-2        38.2kB ± 0%    38.2kB ± 0%   +0.00%  (p=0.029 n=4+4)
Unmarshal/HugoFrontMatter-2          7.44kB ± 0%    7.44kB ± 0%     ~     (all equal)
Marshal/SimpleDocument/struct-2        240B ± 0%      240B ± 0%     ~     (all equal)
Marshal/SimpleDocument/map-2           400B ± 0%      272B ± 0%  -32.00%  (p=0.008 n=5+5)
Marshal/ReferenceFile/struct-2       28.3kB ± 0%    28.3kB ± 0%     ~     (all equal)
Marshal/ReferenceFile/map-2          32.4kB ± 0%    27.9kB ± 0%  -13.82%  (p=0.008 n=5+5)
Marshal/HugoFrontMatter-2            6.45kB ± 0%    5.81kB ± 0%   -9.93%  (p=0.008 n=5+5)
[Geo mean]                           79.8kB         76.7kB        -3.92%

name                               old allocs/op  new allocs/op  delta
UnmarshalDataset/config-2              227k ± 0%      227k ± 0%     ~     (all equal)
UnmarshalDataset/canada-2              782k ± 0%      782k ± 0%     ~     (all equal)
UnmarshalDataset/citm_catalog-2        192k ± 0%      192k ± 0%     ~     (p=0.159 n=5+5)
UnmarshalDataset/twitter-2            56.9k ± 0%     56.9k ± 0%   +0.01%  (p=0.048 n=5+5)
UnmarshalDataset/code-2               1.05M ± 0%     1.05M ± 0%     ~     (all equal)
UnmarshalDataset/example-2            1.36k ± 0%     1.36k ± 0%     ~     (all equal)
Unmarshal/SimpleDocument/struct-2      9.00 ± 0%      9.00 ± 0%     ~     (all equal)
Unmarshal/SimpleDocument/map-2         13.0 ± 0%      13.0 ± 0%     ~     (all equal)
Unmarshal/ReferenceFile/struct-2        183 ± 0%       183 ± 0%     ~     (all equal)
Unmarshal/ReferenceFile/map-2           642 ± 0%       642 ± 0%     ~     (all equal)
Unmarshal/HugoFrontMatter-2             141 ± 0%       141 ± 0%     ~     (all equal)
Marshal/SimpleDocument/struct-2        8.00 ± 0%      8.00 ± 0%     ~     (all equal)
Marshal/SimpleDocument/map-2           10.0 ± 0%       9.0 ± 0%  -10.00%  (p=0.008 n=5+5)
Marshal/ReferenceFile/struct-2          317 ± 0%       317 ± 0%     ~     (all equal)
Marshal/ReferenceFile/map-2             464 ± 0%       429 ± 0%   -7.54%  (p=0.008 n=5+5)
Marshal/HugoFrontMatter-2              90.0 ± 0%      85.0 ± 0%   -5.56%  (p=0.008 n=5+5)
[Geo mean]                            1.12k          1.10k        -1.49%
```